Corrected code for Country and Producer
Updated the blocks to pull in both Country and Producer correcting a variable name in the Producer block.  For country the atoms.py file was updated to recognize that the atom "ctry" maps to country and  the tag import block was updated to pull from the atom first, then seek out alternate tags.